Fix from review part 2 (tests & documentation)
- Clean up small details in code, according to the review

- Improve tests
 - Simplify integration tests
 - split off and merge common QueryableProcessor tests into a separate file
 - move createClassElement function, which is now used in three test files to testutils

- Add documentation to README.md
