+++ README.md
@@ -23,6 +23,7 @@ This package is still in an early phase and the API will likely change.
     1. [Primary Keys](#primary-keys)
     1. [Indices](#indices)
     1. [Ignoring Fields](#ignoring-fields)
+1. [Database Views](#database-views)
 1. [Migrations](#migrations)
 1. [In-Memory Database](#in-memory-database)
 1. [Callback](#callback)
@@ -275,6 +276,9 @@ StreamBuilder<List<Person>>(
 );
 ```
 
+NOTE: It is currently not possible to return a `Stream` if the function queries a database view. This is mostly due 
+to the complexity of detecting which entities are involved in a database view.
+
 ## Transactions
 Whenever you want to perform some operations in a transaction you have to add the `@transaction` annotation to the method.
 It's also required to add the `async` modifier. These methods can only return `Future<void>`.
@@ -414,9 +418,47 @@ class Person {
   Person(this.id, this.name);
 }
 ```
+## Database Views
+If you want to define static `SELECT`-statements which return different types than your entities, your best option is 
+to use `@DatabaseView`. A database view can be understood as a virtual table, which can be queried like a real table.
+
+A database view in floor is defined and used similarly to entities, with the main difference being that
+access is read-only, which means that update, insert and delete functions are not possible. Similarly to 
+entities, the class name is used if no `viewName` was set.
+
+```dart
+@DatabaseView('SELECT distinct(name) as name FROM person', viewName: 'name')
+class Name {
+  final String name;
+
+  Name(this.name);
+}
+```
+
+Database views do not have any foreign/primary keys or indices. Instead, you should manually define indices which fit to 
+your statement and put them into the `@Entity` annotation of the involved entities.
+
+Setters, getters and static fields are automatically ignored (like in entities), you can specify additional fields
+to ignore by annotating them with `@ignore`. 
+
+After defining a database view in your code, you have to add it to your database by adding it to the `views` field of 
+the `@Database` annotation:
+
+```dart
+    @Database(version: 1, entities: [Person], views:[Name])
+    abstract class AppDatabase extends FloorDatabase {
+      PersonDao get personDao;
+    }
+
+```
+
+You can then query the view via a DAO function like an entity.
+
+NOTE: Be aware that it is currently not possible to return a 
+`Stream<>` object from a function which queries a database view.
 
 ## Migrations
-Whenever are doing changes to your entities, you're required to also migrate the old data.
+Whenever you are doing changes to your entities, you're required to also migrate the old data.
 
 First, update your entity.
 Next, Increase the database version.
