+++ floor_generator/lib/value_object/dao.dart
@@ -1,11 +1,10 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:floor_generator/value_object/deletion_method.dart';
+import 'package:floor_generator/value_object/entity.dart';
 import 'package:floor_generator/value_object/insertion_method.dart';
 import 'package:floor_generator/value_object/query_method.dart';
 import 'package:floor_generator/value_object/transaction_method.dart';
 import 'package:floor_generator/value_object/update_method.dart';
-
-import 'entity.dart';
 
 class Dao {
   final ClassElement classElement;
