+++ floor_generator/test/processor/entity_processor_test.dart
@@ -9,9 +9,11 @@ import 'package:floor_generator/value_object/primary_key.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:test/test.dart';
 
+import '../test_utils.dart';
+
 void main() {
   test('Process entity', () async {
-    final classElement = await _createClassElement('''
+    final classElement = await createClassElement('''
       @entity
       class Person {
         @primaryKey
@@ -46,7 +48,7 @@ void main() {
   });
 
   test('Process entity with compound primary key', () async {
-    final classElement = await _createClassElement('''
+    final classElement = await createClassElement('''
       @Entity(primaryKeys: ['id', 'name'])
       class Person {
         final int id;
@@ -77,255 +79,6 @@ void main() {
       constructor,
     );
     expect(actual, equals(expected));
-  });
-
-  test('Ignore hashCode field', () async {
-    final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-      
-        Person(this.id, this.name);
-        
-        @override
-        int get hashCode => id.hashCode ^ name.hashCode;
-      }
-    ''');
-
-    final actual = EntityProcessor(classElement).process();
-
-    expect(actual.fields.length, equals(2));
-  });
-
-  test('Ignore static field', () async {
-    final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-      
-        Person(this.id, this.name);
-        
-        static String foo = 'foo';
-      }
-    ''');
-
-    final actual = EntityProcessor(classElement).process();
-
-    expect(actual.fields.length, equals(2));
-  });
-
-  test('Ignore getter', () async {
-    final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-      
-        String get label => '\$id: \$name'
-      
-        Person(this.id, this.name);
-      }
-    ''');
-
-    final actual = EntityProcessor(classElement)
-        .process()
-        .fields
-        .map((field) => field.name)
-        .toList();
-
-    expect(actual, equals(['id', 'name']));
-  });
-
-  test('Ignore setter', () async {
-    final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-      
-        set printwith(String prefix) => print(prefix+name);
-
-        Person(this.id, this.name);
-        
-        static String foo = 'foo';
-      }
-    ''');
-
-    final actual = EntityProcessor(classElement)
-        .process()
-        .fields
-        .map((field) => field.name)
-        .toList();
-
-    expect(actual, equals(['id', 'name']));
-  });
-
-  group('Constructors', () {
-    test('generate simple constructor', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-      
-        Person(this.id, this.name);
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement).process().constructor;
-
-      const expected = "Person(row['id'] as int, row['name'] as String)";
-      expect(actual, equals(expected));
-    });
-
-    test('generate constructor with named argument', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-        
-        final String bar;
-      
-        Person(this.id, this.name, {this.bar});
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement).process().constructor;
-
-      const expected =
-          "Person(row['id'] as int, row['name'] as String, bar: row['bar'] as String)";
-      expect(actual, equals(expected));
-    });
-
-    test('generate constructor with named arguments', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-        
-        final String bar;
-      
-        Person({this.id, this.name, this.bar});
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement).process().constructor;
-
-      const expected =
-          "Person(id: row['id'] as int, name: row['name'] as String, bar: row['bar'] as String)";
-      expect(actual, equals(expected));
-    });
-
-    test('generate constructor with optional argument', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-        
-        final String bar;
-      
-        Person(this.id, this.name, [this.bar]);
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement).process().constructor;
-
-      const expected =
-          "Person(row['id'] as int, row['name'] as String, row['bar'] as String)";
-      expect(actual, equals(expected));
-    });
-
-    test('generate constructor with optional arguments', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-        
-        final String bar;
-      
-        Person([this.id, this.name, this.bar]);
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement).process().constructor;
-
-      const expected =
-          "Person(row['id'] as int, row['name'] as String, row['bar'] as String)";
-      expect(actual, equals(expected));
-    });
-  });
-
-  group('@Ignore', () {
-    test('ignore field not present in constructor', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-        
-        @ignore
-        String foo;
-      
-        Person(this.id, this.name);
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement)
-          .process()
-          .fields
-          .map((field) => field.name);
-
-      const expected = 'foo';
-      expect(actual, isNot(contains(expected)));
-    });
-
-    test('ignore field present in constructor', () async {
-      final classElement = await _createClassElement('''
-      @entity
-      class Person {
-        @primaryKey
-        final int id;
-      
-        final String name;
-        
-        @ignore
-        String foo;
-      
-        Person(this.id, this.name, [this.foo = 'foo']);
-      }
-    ''');
-
-      final actual = EntityProcessor(classElement).process().constructor;
-
-      const expected = "Person(row['id'] as int, row['name'] as String)";
-      expect(actual, equals(expected));
-    });
   });
 
   group('foreign keys', () {
@@ -377,20 +130,6 @@ void main() {
       expect(actual, equals(expected));
     });
   });
-}
-
-Future<ClassElement> _createClassElement(final String clazz) async {
-  final library = await resolveSource('''
-      library test;
-      
-      import 'package:floor_annotation/floor_annotation.dart';
-      
-      $clazz
-      ''', (resolver) async {
-    return LibraryReader(await resolver.findLibraryByName('test'));
-  });
-
-  return library.classes.first;
 }
 
 Future<List<ClassElement>> _createClassElements(final String classes) async {
